Move `SimulationConfig(Run)` to `simulation_structure`
