+++ packages/engine/lib/simulation_structure/src/simulation/config.rs
@@ -4,18 +4,42 @@ use execution::{
     package::simulation::{PackageCreatorConfig, PersistenceConfig, SimulationId},
     worker_pool::WorkerAllocation,
 };
-use simulation_structure::ExperimentConfig;
 use stateful::{field::Schema, global::Globals, state::StateCreateParameters};
 
-pub use self::{
-    error::{Error, Result},
-    simulation::SimulationConfig,
-};
+use crate::ExperimentConfig;
 
-mod error;
-mod simulation;
+const MIN_AGENTS_PER_GROUP: usize = 10;
 
-pub const MIN_AGENTS_PER_GROUP: usize = 10;
+pub struct SimulationConfig {
+    pub id: SimulationId,
+    pub schema: Arc<Schema>,
+    pub worker_allocation: Arc<WorkerAllocation>,
+    pub max_num_steps: usize,
+    pub package_creator: PackageCreatorConfig,
+}
+
+impl SimulationConfig {
+    pub fn new(
+        id: SimulationId,
+        globals: Globals,
+        worker_allocation: WorkerAllocation,
+        schema: Schema,
+        persistence_config: PersistenceConfig,
+        max_num_steps: usize,
+    ) -> Self {
+        Self {
+            id,
+            package_creator: PackageCreatorConfig {
+                agent_schema: Arc::clone(&schema.agent_schema),
+                globals,
+                persistence: persistence_config,
+            },
+            worker_allocation: Arc::new(worker_allocation),
+            schema: Arc::new(schema),
+            max_num_steps,
+        }
+    }
+}
 
 pub struct SimulationRunConfig {
     experiment: Arc<ExperimentConfig>,
@@ -32,7 +56,7 @@ impl SimulationRunConfig {
         persistence_config: PersistenceConfig,
         max_num_steps: usize,
     ) -> SimulationRunConfig {
-        let simulation_config = simulation_config(
+        let simulation_config = SimulationConfig::new(
             id,
             globals,
             worker_allocation,
@@ -62,26 +86,5 @@ impl SimulationRunConfig {
             agent_schema: Arc::clone(&self.simulation.schema.agent_schema),
             message_schema: Arc::clone(&self.simulation.schema.message_schema),
         }
-    }
-}
-
-fn simulation_config(
-    id: SimulationId,
-    globals: Globals,
-    worker_allocation: WorkerAllocation,
-    schema: Schema,
-    persistence_config: PersistenceConfig,
-    max_num_steps: usize,
-) -> SimulationConfig {
-    SimulationConfig {
-        id,
-        package_creator: PackageCreatorConfig {
-            agent_schema: Arc::clone(&schema.agent_schema),
-            globals,
-            persistence: persistence_config,
-        },
-        worker_allocation: Arc::new(worker_allocation),
-        schema: Arc::new(schema),
-        max_num_steps,
     }
 }
