+++ packages/engine/lib/simulation_structure/src/simulation/mod.rs
@@ -1,10 +1,14 @@
+mod config;
 mod package_creators;
 
 use execution::package::simulation::PackageInitConfig;
 use serde::{Deserialize, Serialize};
 use stateful::global::Dataset;
 
-pub use self::package_creators::PackageCreators;
+pub use self::{
+    config::{SimulationConfig, SimulationRunConfig},
+    package_creators::PackageCreators,
+};
 
 /// Analogous to `SimulationSrc` in the web editor
 /// This contains all of the source code for a specific simulation, including
