--- packages/engine/src/config/simulation.rs
@@ -1,15 +1,0 @@
-use std::sync::Arc;
-
-use execution::{
-    package::simulation::{PackageCreatorConfig, SimulationId},
-    worker_pool::WorkerAllocation,
-};
-use stateful::field::Schema;
-
-pub struct SimulationConfig {
-    pub id: SimulationId,
-    pub schema: Arc<Schema>,
-    pub worker_allocation: Arc<WorkerAllocation>,
-    pub max_num_steps: usize,
-    pub package_creator: PackageCreatorConfig,
-}
