+++ packages/engine/src/lib.rs
@@ -35,14 +35,12 @@
 )]
 
 mod args;
-pub mod config;
 pub mod datastore;
 pub mod env;
 mod error;
 pub mod experiment;
 pub mod proto;
 pub mod simulation;
-pub mod types;
 pub mod utils;
 
 pub use self::{
