+++ packages/engine/src/simulation/package/run.rs
@@ -13,7 +13,7 @@ use execution::{
 };
 use futures::{executor::block_on, stream::FuturesOrdered, StreamExt};
 use memory::shared_memory::MemoryId;
-use simulation_structure::PackageCreators;
+use simulation_structure::{PackageCreators, SimulationRunConfig};
 use stateful::{
     context::{Context, ContextColumn, PreContext},
     field::{FieldSource, FieldSpecMapAccessor},
@@ -21,10 +21,7 @@ use stateful::{
 };
 use tracing::{Instrument, Span};
 
-use crate::{
-    config::SimulationRunConfig,
-    simulation::error::{Error, Result},
-};
+use crate::simulation::error::{Error, Result};
 
 /// Represents the packages of a simulation engine.
 pub struct Packages {
