--- packages/engine/src/types.rs
@@ -1,4 +1,0 @@
-// TODO: UNUSED: Needs triage
-pub const BATCH_ID_LENGTH: usize = 32;
-// TODO: UNUSED: Needs triage
-pub type BatchId = [u8; BATCH_ID_LENGTH];
